v 0.2.2.a - Data exploder, minor bugfixes & README
This version brings a new data editing module to the table. Ever needed to chop one file with multiple sets of data exploded into multiple files? Just turn on the "data exploder" module and pull a Kerbal Space Program on the file!

Also, there is a new README file, just waiting to be completed! This commit introduces some information on how to install the software and it's dependencies.

Bug fixes:
-> If labels are limited by ' " ' sign, the label separator sign will be ignored during preparation of returnable labels in projectexporter.readData() function,
-> Default graphing engine is set to "modules/engine.py",
-> Fixed horrible spelling in change log.