v. 0.2.4 - Direct legend editing
This update introduces direct data editing in Legend Tab, some changes in project exporter (to accomodate new variables available) and linewidth and alpha handling in engine.py.
Also, I had to edit some of the most atrocious spelling and grammar mistakes I've ever done publicly, both in the CHANGELOG and in commit description, as well as some minor changes to the graphical side of the file.